requirements.txt
FastAPI==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
pandas==2.3.3
numpy>=1.26.0
beautifulsoup4==4.12.2
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
torch>=2.0.0
python-dotenv==1.0.0
pydantic==2.5.2
